requirements.txt
dlib
numpy
opencv-python
pandas
Pillow
pydantic
python-dateutil
pytz
six
typing_extensions
